+++ src/js/main.js
@@ -6,19 +6,21 @@ import { Provider } from 'react-redux'
 import { IndexRoute, Route, Router } from 'react-router'
 import { combineReducers, compose, createStore } from 'redux'
 import { reduxReactRouter, routerStateReducer } from 'redux-router'
+import history from './helpers/history'
+import { bookmarks, endOfBookmarks, networkState, page, searchTerm, toaster } from './helpers/reducers'
 import { createHistory } from 'history'
-import { bookmarks, endOfBookmarks, networkState, page, toaster } from './helpers/reducers'
-import createBrowserHistory from 'history/lib/createBrowserHistory'
 import io from 'socket.io-client'
 import Bkmrkd from './containers/bkmrkd'
 import Colophon from './components/colophon'
 import Bookmarks from './components/bookmarks'
 import { addToast } from './helpers/actions'
+import Search from './components/search'
 
 const routes = (
   <Route path='/' component={Bkmrkd}>
     <IndexRoute component={Bookmarks} />
     <Route path='/colophon' component={Colophon} />
+    <Route path='/search' component={Search} />
   </Route>
 )
 
@@ -30,7 +32,8 @@ if (typeof window !== 'undefined') {
     networkState: '',
     toaster: [],
     page: 1,
-    endOfBookmarks: false
+    endOfBookmarks: false,
+    searchTerm: ''
   }
 
   const reducer = combineReducers({
@@ -39,7 +42,8 @@ if (typeof window !== 'undefined') {
     networkState,
     toaster,
     page,
-    endOfBookmarks
+    endOfBookmarks,
+    searchTerm
   })
   const store = compose(
     reduxReactRouter({
@@ -97,7 +101,7 @@ if (typeof window !== 'undefined') {
     }))
   })
 
-  render(<Provider store={store}><Router history={createBrowserHistory()}>{ routes }</Router></Provider>, document.body.querySelector('[data-hook="app"]'))
+  render(<Provider store={store}><Router history={history}>{ routes }</Router></Provider>, document.body.querySelector('[data-hook="app"]'))
 }
 
 export default routes