+++ test/seeds/data.js
@@ -1,0 +1,134 @@
+import rethink from 'rethinkdb'
+
+export default [
+  {
+    title: 'test 1',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 2',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 3',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 4',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 5',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 6',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 7',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 8',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 9',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 10',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 11',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 12',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 13',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 14',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 15',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 16',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 17',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 18',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 19',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 20',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 21',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 22',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 23',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 24',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 25',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  },
+  {
+    title: 'test 26',
+    url: 'https://duckduckgo.com',
+    createdOn: rethink.now()
+  }
+]